Flask app using a machine learning model to predict life insurance application response.
A machine learning model was built with the data from a kaggle competition where the response to a life insurance application is to be predicted.
Since the data is that of real customers of Prudential Life Insurance, it is altered to be anonymous with the response being designated as 1 to 8.